Added task scheduler and got regular scheduled Pig jobs sort of working... needs polish and ideally can replace current hack fix for Java Pig API issues

+++ event_receiver/src/main/java/com/cloudrecommend/StreamProcessor.java
@@ -34,7 +34,7 @@ public class StreamProcessor implements Runnable {
         this.hdfsUri = "hdfs://" + hdfsUri + "/";
 
         // unique ID for this process and machine
-        id = Long.toString(System.currentTimeMillis()) + Double.toString(Math.round((Math.random() * 100))) + processId;
+        id = Long.toString(System.currentTimeMillis()) + Long.toString((int)Math.round((Math.random() * 100))) + processId;
     }
 
     @Override
